2004-10-28 Erik Hjortsberg <erik@hysteriskt.nu> http://erikhjortsberg.blogspot.com/
* moved to Eris 1.3
This requires some changes to the way enties are handled etc..
There's a big bug which I can't seem to track down. It manifests itself in that entities attached to other entities (i.e. not directly attached to the root node) can't change their orientation. An interim solution for now is to direct all entities attached to the world to the root node. This works for now, but the effect can be seen when the avatar walk into the sty.
*moved to Ogre 0.15.0
Ember now compiles and works against the latest ogre point release. We'll try to keep it this way.
As a result of the move to 0.15.0 I've updated the meshes (Since there's been changes to the binary .mesh format. New meshes can be downloaded from http://purple.worldforge.org/~erik/ember/media/Media.tar.gz.).
*added some more functionality to ServerBrowserWidget and ServerWidget