Fixed terrain loading which was broken by Ogre 1.4.8.
With 1.4.8, additional checks have been added when a file is loaded from the filesystem through the Ogre resource system. These checks make sure that if an absolute file name is used, it still must reside somewhere in the predefined resource locations.
The terrain.cfg file will however not reside in any location included in the ogre resource paths, and thus the Ogre resource system can't be used. We've therefore modified the PLSM2 to be able to load all options through a stream object instead . We create this stream object ourselves and never touches the Ogre system.