Cleaned up the LoggingService header and made sure that we have a logging instance all the time.
The LoggingService isn't handled like other services since we depend on it being there for all of our loggin needs. We really should refactor the logging methods to be static instead of being dependent on a service. The static methods could very well interface with a service, but there should never be a catastrophic segfault if it wasn't created when we want to log.