Don't try to use a common OgreIteratorWrapper.pkg file with nested templates, as this will only confuse tolua and prevent it from properly garbage collect entries that are iterated over.
The end result was that when iterating over any subclass of Ogre::ResourcePtr, the pointers would never be freed. Since the pointers uses reference counting and only destroy themselves when nothing is holding a reference, that meant that no resource which ever had been iterated over in the gui was never destroyed.