Update the foliage when the terrain is altered.
We would like this to only update the foliage areas that are actively affected by the terrain change, but I couldn't get that to work correctly, so we instead update the whole foliage. Newer versions of PagedGeometry has better methods for this, so we'll wait until we've updated to that.