Provide our own shadow pass setup method.
The reason we want to do this is because we want to reuse the current
pass' settings for the shadow caster pass, even when using technique
specific caster materials. The default method will however only do this
when the SceneManager wide custom shadow caster material is set.
By providing this we can much easier define different shadow caster
materials for each technique.