Added lights to the model as well as entities and particle systems.
This means that a Model instance is now completly self contained, so
there's no need to set up the attachment of the lights or particle
systems separately. It also means that the ability to use a position
offset for these is lost. If there's a need for having an offset (either
position or orientation) we would need to alter this so that the lights
and particle systems somehow are separated from the entities.