Made the connection between the Ogre scene manager and the TerrainManager more loose, so to avoid race conditions when bridges are deleted.
The bridges are now registered with the TerrainManager, where they will
be stored using reference counting. This will ensure that they won't be
deleted as long as at least one thread is referencing them.
The same now also applies to the height data, which will also be
reference counted. This will deal with the situation where a page is
unloaded while it's height data is being processed in a background
thread.

Also removed most of the Ogre code from TerrainManager and put it in
TerrainPageDataProvider. There are still some Ogre hooks left though.