Altered the way the lua bindings are handled.
Instead of providing the bindings in the source code we now only provide
the .pkg files and depend on the bindings being generated at build time.
The main advantages of this is that it fixes some issues with different
versions of tolua++ producing different code.
This adds a new build time requirement of having the tolua++ executable
available though.