Refactor terrain mod handling.
To use the new API we need to separate out the parsing of the atlas data
to the background tasks. Since this is too tightly coupled to the
main Eris polling loop in Eris::TerrainMod we won't use that class at
all, instead directly using InnerTerrainMod. The handling of all of the
terrain data then occurs in the TerrainMod*Task classes.