Keep separate configs for different scopes.
In order to better be able to correctly save changed configs, we need to
keep the config scopes separate. When the app is requested to save
configuration, it should then just save those values that either have
changed (i.e. instance scope, if different from global scope) or those
values that already were defined in the user scope (i.e. written to the
config file).