Fixed issue with comparing floating values.
Due to the way floating values are represented in binary it's not always
possible to correctly say whether two floating point values are the
same. Rather than having to fiddle with epsilon values we'll just
convert the values to strings and compare them.

This fixes an issue in the settings widget where some settings always
were considered changed.