Ignore autogenerated binding files.
Though a better way to handle this is to always build in a directory
separate from the source. If this directory is a child directory of the
source directory, you can edit .git/info/exclude to add any repo
specific exclude rules there.

Though I'll add an ignore for these files here anyway, since I suspect
that not all will follow these guidelines.