Reduce usage of singletons and optimize foliage density updates.
Constructors are used to pass references the required references instead of using singletons to obtain the references. A further change while removing singleton calls was to move listening for the foliage density updates from the GrassFoliage to the Foliage class itself. Therefore, instead of every foliage layer listening for an update, only the foliage class listens for the updates and then iterates and updates all the foliage.