Handle extra libs in boost 1.50.0.
With Boost 1.50.0 the boost_thread library also depends on the
boost_chrono and boost_system lib. To handle this we've added a check
whcih uses "ldd" to see what libraries of boost are needed.

Also added an extra check for boost libraries in the optionally
"--with-boost" path (since these libraries should be searched for first,
before falling back to system check).