Added composition mode.
This allows the specification of some entities to be "composition"
entities, which means that they together with their child entities
constitute one conceptual entity.
We provide three different kind of composition modes:
* none: no composition; this is the default
* composition: When the user interacts with a child entity of
the composition she or he should be presented with the option to also
interact with the composition entity.
* exclusive composition: When the user interacts with a child entity of
the composition she or he should only presented with the composition
entity.