Handle attaching entities without Models.
We'll have to do this in a pretty hackish way current, through the
creation of a temporary Ogre::ManualObject which is then destroyed. This
is not optimal.
One other option would be to extend Model to allow for attachment
without having to specify a MovableObject. Another approach would be to
bake the scene nodes into static geometry first.