Provide workaround for OpenGL crash.
The issue is that if the index list is empty, the OpenGL renderer will
segfault (Direct3d is fine). Our hacky fix is to always make sure that
there's at least one triangle available. We do this by specifying an
"empty" triangle where all vertices are the same.
Ogre should be patched to better handle this, and this hack removed.