Use int for characters.
That way we'll cover the whole UTF range as provided by SDL.
This if course relies on the assumption that int resolves to at least 16
bits on the target machine. I think we can be pretty secure in knowledge
that any modern machine which will run Ember will also use at least 16
bits for int though.