Fixed issue with pages being unloaded.
This was caused by the loadedList containing duplicate page entries.
This resulted in some pages being unloaded when the geometry was
updated. The reason for this was the changes we've made before which
allowed for background loading (which could result in duplicate list
entries).
The solution was to use a "set" instead. This might have some
performance drawbacks when pages are added, but as this happens so
seldomly the effect should be negligible.