Add class for local admin login.
When connecting to a local socket the user has the option to log in as
an admin client even though no password or likewise is known. This class
helps with the flow, hooking into the various steps the server service
goes through to both create an account and to log in the user as a
"creator" character.