Refer to goals through their ids.
When interacting with specific goals these are now referred to by their
id. This is obtained by first sending a Commune op with "goal" specified
as an attribute.
This way is much more preferrable to having to involve the "subject",
which is an Interlinguish specific construct which doesn't really have
any relevance to direct goal manipulation.