Send arguments on to the executable.
Since the scripts executes commands the arguments were cleared. That's
fixed by stashing them into "orig_args" first.
Secondly the usage of quotes for the $orig_args made them not appear to
the executable. The quotes were removed. It's not clear what the reason
for adding them in the first place was.