Don't start paging until avatar entered world.
Before the avatar has entered the world the camera is placed at origo.
However, once it's entered into the world the correct position is used.
The problem with starting the terrain paging too early is that the
background queue will then be filled up with requests for creating pages
that aren't near the avatar (since the avatar might not be near origo),
causing delays since these must be processed before the pages that
actually are near the avatar are handled.