Added fix for GL context loss bug.
When running KDE 4.14.4 with OpenGL enabled in the windowmanager (kwin)
it seems that when minimizing and restoring the window the GL context
sometimes gets altered. The result is a segfault when Ogre tries to
issue GL commands.
The bug is stochastic, but can be provoked by minimizing the window and
restoring it a couple of times. The exact cause is unknown, but it
appears that the bug is avoided by making sure that the GL context is
set each time the window gets altered. However, there's no guarantee;
I've just observed that after this fix is applied I haven't yet been
able to trigger the bug.
This should of course be fixed in the faulty component (which I think is
KWin, but might also be x.org (version 1.15.1) or the Nvidia drivers
(337.12)).