Generate mesh tangents if missing.
The tangents can in almost all cases be generated from the existing mesh
data. But it takes some time, so it's preferred to have them
pregenerated.
By making this automatic we can easier work with development media. It's
expected that all release media should be processed, so tangents should
always be generated for that. If not, we'll log with a warning.