Fixed issue with incorrect vertex data.
This bug took quite some time to track down.
It appears that on at least gcc 4.8.3 there's some kind of bug which
makes the value assigned be off by one. I.e. if the map is empty, and
we're inserting the first instance, the "size()" method would return
"0", and that's the value we would be inserting. However, for some
reason which might have to do with caching, optimization or out-of-order
execution, the value written is actually the size _after_ the insertion.
So if the map was empty, the first entry would get the value of "1",
which would be the size _after_ the insertion.
And this would of course cause issues later on, with the memcpy writing
beyond the data, thus corrupting the memory, leading to segfaults.