Handle minds by checking type, and entity refs.
Since attributes unchanged from the type aren't sent we need to check
with the type if the entity has a mind.
Also implemented parsing of entity references in the mind code. An
entity is referenced using the "$eid:" syntax. These references needs to
be adjusted with the new id values of the created entities.