Added optional feature analyzing dropped frames.
This relies on the unwind library, which helps with printing out a
stack trace.
The idea here is to give some information about what's happening when a
frame takes too long. While there are multiple reasons for these
occurences, one helpful hint would be to get a sample of what the main
thread is doing when its taking too long.

This feature will create a background thread, which will detect when the
main thread has taken too long, and then send a signal to it. The signal
will be executed on the main thread and will use the unwind library to
create a backtrace, which will be printed to the console.
In some situations this might be helpful for a developer in figuring out
why a frame took too long.