Check for WF libraries first.
The reasoning is that these libraries most often are built by the user,
and are not system libraries. By checking for them first we try to make
sure that locally installed paths are included firsts in the list of
includes (i.e. "-L"), which would prevent any system installed WF
libraries to be picked instead of locally installed ones.