Let the parent attachment decide how to represent the entity.
We want different representations depending on the parent.
The "physical" domain entity should result in graphical representations. Other parents
should result in hidden entities.
The deviation is entities that are attached. In those case we'll show a model directly, attached to
a bone.
The idea is that some entities that are attached should invoke "effects" in their parent entities.
The typical example would be the a fire, when contained in another entity should activate the
"fire effect" on the parent entity, enveloping it in flames.