Fixed issue where nothing was rendered if the cursor was disabled.
The reason seemed to be that the last rendered window in that situation set the scissor test to a portion of the screen,
and Ogre didn't reset it.
With the cursor this wasn't an issue, since it's always rendered last and it sets the scissor test to the whole screen.