Use HW Instance material.
Where individual mesh data is stored in extra UV coords. This is more efficient
than the shader based approach.
The issues with PSSM shadows we experienced before is gone. It's unclear if it's
fixed in Ogre, or if it was an effect of incorrect LODs (which is fixed in Ogre 1.12.13).