Remove samplers use in core scripts.
This is because of a peculiar issue in Ogre where samplers aren't unloaded
along with the resource group. To handle this we unload _all_ samplers when we also unload
the "world" resource group. As a result of this however we cannot use any samplers at all
in our core scripts (as they would then also be unloaded).